starter-bundles/mcp-server/requirements.txt
fastmcp
httpx
uvicorn
mcpo
